requirements.txt
django==1.10.6
requests
Pillow
bs4
lxml
gunicorn
transliterate
pytz